Modify

+++ src/main/java/service/OrderService.java
@@ -35,22 +35,24 @@ public class OrderService {
             switch (choice) {
                 case 1:
                     if (isAdmin(currentMember)) {
-                        allOrderSelect();
-                    }else {
-                        createOrder(cartSession, currentMember);
+                        viewAllOrders();
+                    } else {
+                        if (cartSession.getCart().isEmpty()) {
+                            System.out.println("장바구니가 비어있습니다.");
+                        } else {
+                            createOrder(cartSession, currentMember);
+                        }
                     }
                     break;
                 case 2:
                     if (isAdmin(currentMember)){
                         changeOrderStatus(sc);
-                    }else {
+                    } else {
                         findOrders(sc);
                     }
                     break;
                 case 3:
-                    if (!isAdmin(currentMember)){
-                        deleteOrder(sc, currentMember);
-                    }
+                    if (!isAdmin(currentMember)) deleteOrder(sc, currentMember);
                     break;
                 case 0:
                     return;
@@ -87,15 +89,18 @@ public class OrderService {
     }
 
     private void findOrders(Scanner sc) {
+        /**
+         * 본인의 주문 목록 조회
+         */
         Long memberId = Session.getInstance()
                 .getCurrentMember()
                 .getMemberId();
-        List<Order> orderList = orderRepository.findByMemberId(memberId);
-        if (orderList.isEmpty()) {
+        List<Order> orders = orderRepository.findByMemberId(memberId);
+        if (orders.isEmpty()) {
             System.out.println("주문 내역이 없습니다.");
             return;
         } else {
-            orderList.forEach(o -> {
+            orders.forEach(o -> {
                 System.out.println("주문 ID: " + o.getOrderId()
                         + " | 주문날짜: " + o.getDate()
                         + " | 총 주문 금액: " + o.getTotalPrice()
@@ -104,29 +109,61 @@ public class OrderService {
             );
         }
 
-        Long id = inputLong("상세 조회할 주문 번호 : ", sc);
-        List<OrderItem> orderItems = orderItemRepository.findByOrderId(id);
+        //todo: viewOrderDetails, createItemReview 메서드 추출
 
-        if (orderItems.isEmpty()){
-            System.out.println("주문 번호를 확인해주세요.");
-        }else {
-            orderItems.stream()
-                    .forEach(o -> {
-                        System.out.println("상품명: " + extractItemName(o.getItemId())
-                                + " | 수량: " + o.getQuantity()
-                                + " | 금액: " + o.getPrice());
-                    });
-//            Long itemId = inputLong("리뷰 작성할 상품 번호 선택 : ",sc);
-//
-//            for (int i = 0; i < orderItems.size(); i++) {
-//                if (itemId.equals(orderItems.get(i).getItemId())) {
-//                    int stars = inputInt("평점을 입력해주세요: ", sc);
-//                    String contents = inputString("내용을 입력해주세요: ", sc);
-//                    Review review = Review.of(stars,contents,memberId,itemId);
-//                    reviewRepository.save(review);
-//                }
-//            } // 생각나는대로 작성해서 조금 더 생각해보겠슴다
+        /**
+         * 본인 주문 상세 조회
+         */
+        Long id = inputLong("상세 조회할 주문 번호 선택 : ", sc);
+        if (!validateOrderId(orders, id)){
+            System.out.println("없는 주문이거나 본인의 주문이 아닙니다.");
+            return;
         }
+
+        List<OrderItem> orderItems = orderItemRepository.findByOrderId(id);
+        orderItems.stream()
+                .forEach(oi -> {
+                    System.out.println("상품 ID: " + oi.getItemId()
+                            + " | 상품명: " + extractItemName(oi.getItemId())
+                            + " | 수량: " + oi.getQuantity()
+                            + " | 금액: " + oi.getPrice());
+                });
+
+        orderItemRepository.findByOrderId(id)
+                .stream()
+                .filter(oi -> oi.getItemId().equals(id))
+                .map(oi -> "상품 ID: " + oi.getItemId()
+                        + " | 상품명: " + extractItemName(oi.getItemId())
+                        + " | 수량: " + oi.getQuantity()
+                        + " | 금액: " + oi.getPrice())
+                .findAny()
+                .ifPresent(System.out::println);
+
+        /**
+         * 특정 주문의 상품에 대한 리뷰 작성
+         */
+        Long itemId = inputLong("리뷰 작성할 상품 번호 선택 : ", sc);
+
+        orderItems.stream()
+                .filter(oi -> itemId.equals(oi.getItemId()))
+                .map(oi -> createReview(memberId, itemId, sc))
+                .findAny()
+                .ifPresent(reviewRepository::save);
+
+//        for (int i = 0; i < orderItems.size(); i++) {
+//            if (itemId.equals(orderItems.get(i).getItemId())) {
+//                int stars = inputInt("평점을 입력해주세요: ", sc);
+//                String contents = inputString("내용을 입력해주세요: ", sc);
+//                Review review = Review.of(stars, contents, memberId, itemId);
+//                reviewRepository.save(review);
+//            }
+//        }
+    }
+
+    //주문 ID 검증
+    private static boolean validateOrderId(List<Order> orders, Long id) {
+        return orders.stream()
+                .anyMatch(o -> o.getOrderId().equals(id));
     }
 
     private String extractItemName(Long itemId){
@@ -134,13 +171,19 @@ public class OrderService {
         return item.getName();
     }
 
+    private Review createReview(Long memberId, Long itemId, Scanner sc) {
+        int stars = inputInt("평점을 입력해주세요: ", sc);
+        String contents = inputLine("내용을 입력해주세요: ", sc);
+        return Review.of(stars, contents, memberId, itemId);
+    }
+
+    //== 주문 취소 ==//
     private void deleteOrder(Scanner sc, Member currentMember) {
         Long id = inputLong("취소할 주문 번호: ", sc);
         Order order = orderRepository.findById(id).orElse(null);
         if (order == null) {
-            System.out.println("그런 주문은 없습니다.");
+            System.out.println("없는 주문 번호 입니다.");
         } else {
-
             if (order.getMemberId().equals(currentMember.getMemberId()) && checkOrderStatus(order)) {
                 orderItemRepository.deleteByOrderId(id);
                 orderRepository.deleteById(id);
@@ -157,14 +200,23 @@ public class OrderService {
         return currentMember.getRole().equals(Role.ADMIN);
     }
 
+    //== 주문 상태 변경 ==//
     private void changeOrderStatus(Scanner sc) {
         Long id = inputLong("변경할 주문 번호: ", sc);
+        Optional<Order> order = orderRepository.findById(id);
+        if (!order.isPresent()) {
+            System.out.println("주문 번호를 다시 확인해주세요.");
+            return;
+        }
 
         OrderStatus orderStatus = null;
         while (orderStatus == null) {
             try {
-                int statusNum = inputInt("변경할 상태(" + Arrays.toString(OrderStatus.values()) + ')', sc);
-                orderStatus = getOrderStatus(statusNum);
+                int statusNum = inputInt("변경할 상태("
+                        + Stream.of(OrderStatus.values())
+                        .map(os -> os.name() + ":" + os.getStatusNum())
+                        .collect(Collectors.joining(", ")) + "): ", sc);
+                orderStatus = OrderStatus.fromStatusNum(statusNum);
             } catch (IllegalArgumentException e) {
                 System.out.println("잘못된 입력입니다.");
             }
@@ -175,20 +227,15 @@ public class OrderService {
         System.out.println("주문 번호 " + id + "의 상태가 " + orderStatus + "로 변경되었습니다.");
     }
 
-    private static OrderStatus getOrderStatus(int statusNum) {
-        for (OrderStatus status : OrderStatus.values()) {
-            if (status.getStatusNum() == statusNum) {
-                return status;
-            }
-        }
-        return null;
-    }
-
-    private void allOrderSelect() {
+    private void viewAllOrders() {
         List<Order> orderList = orderRepository.findAll();
-        for (Order order : orderList) {
-            System.out.println(order);
-        }
+        orderList.forEach(o -> {
+            System.out.println("주문 ID: " + o.getOrderId()
+                    + " | 주문날짜: " + o.getDate()
+                    + " | 총 주문 금액: " + o.getTotalPrice()
+                    + " | 주문상태: " + o.getStatus()
+            );}
+        );
     }
 
     private void displayOrderMenu() {
